+++ R/watercomp.R
@@ -10,8 +10,7 @@ mwlSource = function(obs, MWL=c(8.01, 9.57, -8.096, 2564532.2, 5.76, 80672),
   if(MWL[6] == 80672 & stype == 2){
     warning("Using stype=2 and GMWL is inappropriate for most applications; see man")
   }
-  
-  if(length(class(obs)) < 2 | class(obs)[2] != "iso"){
+  if(!(inherits(obs, "iso"))){
     warning("Expecting iso object for obs, this argument may be formatted incorrectly")
   }
   if(!(stype %in% c(1, 2))){
@@ -32,7 +31,7 @@ mwlSource = function(obs, MWL=c(8.01, 9.57, -8.096, 2564532.2, 5.76, 80672),
   
   #establish credible range for source water d18O
   #center calculated as intersection of EL and MWL
-  o_cent = (MWL[2] - (obs$H - slope[1] * obs$O) ) / (slope[1]-MWL[1])
+  o_cent = (MWL[2] - (mean(obs$H) - slope[1] * mean(obs$O)) ) / (slope[1]-MWL[1])
   
   #large range of d18O values to evaluate
   o_eval = seq(o_cent - 25, o_cent + 25, by = 0.001)
@@ -101,9 +100,8 @@ mwlSource = function(obs, MWL=c(8.01, 9.57, -8.096, 2564532.2, 5.76, 80672),
 mixSource = function(obs, sources, slope, prior=rep(1,nrow(sources)), 
                    shp=1, ngens=1e5, ncores = 1){
 
-  if(length(class(obs)) < 2 | class(obs)[2] != "iso"){
-    warning("Expecting iso object for obs, this argument may be
-            formatted incorrectly")
+  if(!(inherits(obs, "iso"))){
+    warning("Expecting iso object for obs, this argument may be formatted incorrectly")
   }
   
   #get number of observations